Rename swap -> spl-token-swap and serum-swap -> swap
