+++ packages/spl-token-swap/package.json
@@ -1,0 +1,49 @@
+{
+  "name": "@project-serum/spl-token-swap",
+  "version": "0.0.14",
+  "description": "Serum Swap",
+  "main": "lib/index.js",
+  "types": "lib/src/index.d.ts",
+  "license": "Apache-2.0",
+  "publishConfig": {
+    "access": "public"
+  },
+  "scripts": {
+    "build": "tsc",
+    "test": "jest test",
+    "coverage": "jest --coverage test",
+    "prepublishOnly": "yarn build",
+    "shell": "tsc && node -e \"$(< shell)\" -i --experimental-repl-await"
+  },
+  "jest": {
+    "transform": {
+      ".(ts)": "ts-jest"
+    },
+    "testEnvironment": "node",
+    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts)$",
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js"
+    ]
+  },
+  "dependencies": {
+    "@project-serum/serum": "^0.13.21",
+    "@solana/spl-token": "^0.0.13",
+    "@solana/spl-token-swap": "0.1.0",
+    "bn.js": "^5.1.3",
+    "bs58": "^4.0.1",
+    "buffer-layout": "^1.2.0",
+    "dotenv": "^8.2.0"
+  },
+  "peerDependencies": {
+    "@solana/web3.js": "^0.90.0"
+  },
+  "devDependencies": {
+    "@tsconfig/node12": "^1.0.7",
+    "@types/bn.js": "^4.11.6",
+    "@types/bs58": "^4.0.1",
+    "@types/jest": "^24.9.1",
+    "@types/node": "^12.12.62"
+  }
+}
