+++ packages/spl-token-swap/src/index.ts
@@ -1,0 +1,4 @@
+export * from './pools';
+export * from './instructions';
+export * from './utils';
+export * from './types';
