+++ packages/spl-token-swap/src/utils.ts
@@ -1,0 +1,16 @@
+import BN from 'bn.js';
+
+export const timeMs = (): number => {
+  return new Date().getTime();
+};
+
+export function divideBnToNumber(numerator: BN, denominator: BN): number {
+  const quotient = numerator.div(denominator).toNumber();
+  const rem = numerator.umod(denominator);
+  const gcd = rem.gcd(denominator);
+  return quotient + rem.div(gcd).toNumber() / denominator.div(gcd).toNumber();
+}
+
+export function getTokenMultiplierFromDecimals(decimals: number): BN {
+  return new BN(10).pow(new BN(decimals));
+}
