+++ packages/swap/README.md
@@ -2,7 +2,7 @@
 
 [![npm (scoped)](https://img.shields.io/npm/v/@project-serum/serum-swap)](https://www.npmjs.com/package/@project-serum/serum-swap)
 [![Build Status](https://travis-ci.com/project-serum/serum-ts.svg?branch=master)](https://travis-ci.com/project-serum/serum-ts)
-[![Documentation](https://img.shields.io/badge/typedoc-documentation-blue)](https://project-serum.github.io/serum-ts/serum-swap/index.html)
+[![Documentation](https://img.shields.io/badge/typedoc-documentation-blue)](https://project-serum.github.io/serum-ts/serum-swap/classes/swap.html)
 [![Discord Chat](https://img.shields.io/discord/739225212658122886?color=blueviolet)](https://discord.com/channels/739225212658122886)
 [![License](https://img.shields.io/github/license/project-serum/serum-dex?color=blue)](https://opensource.org/licenses/Apache-2.0)
 
@@ -14,7 +14,7 @@ Using npm:
 
 ```
 npm install @solana/web3.js @project-serum/serum-swap
-```
+```cd
 
 Using yarn:
 
@@ -24,4 +24,4 @@ yarn add @solana/web3.js @project-serum/serum-swap
 
 ## API Reference
 
-[API Reference](https://project-serum.github.io/serum-ts/serum-swap/index.html).
+[API Reference](https://project-serum.github.io/serum-ts/serum-swap/classes/swap.html).
