+++ packages/swap/package.json
@@ -1,49 +1,65 @@
 {
   "name": "@project-serum/swap",
-  "version": "0.0.16",
-  "description": "Serum Swap",
+  "version": "0.1.0-alpha.1",
+  "description": "Client for swapping on the serum dex",
+  "license": "MIT",
+  "repository": "project-serum/serum-ts",
   "main": "lib/index.js",
+  "source": "src/index.js",
   "types": "lib/index.d.ts",
-  "license": "Apache-2.0",
-  "publishConfig": {
-    "access": "public"
+  "engines": {
+    "node": ">=10"
   },
   "scripts": {
     "build": "tsc",
-    "test": "jest test",
-    "coverage": "jest --coverage test",
-    "prepublishOnly": "yarn build",
-    "shell": "tsc && node -e \"$(< shell)\" -i --experimental-repl-await"
-  },
-  "jest": {
-    "transform": {
-      ".(ts)": "ts-jest"
-    },
-    "testEnvironment": "node",
-    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts)$",
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js"
-    ]
-  },
-  "dependencies": {
-    "@project-serum/serum": "^0.13.21",
-    "@solana/spl-token": "^0.0.13",
-    "@solana/spl-token-swap": "0.1.0",
-    "bn.js": "^5.1.3",
-    "bs58": "^4.0.1",
-    "buffer-layout": "^1.2.0",
-    "dotenv": "^8.2.0"
-  },
-  "peerDependencies": {
-    "@solana/web3.js": "^0.90.0"
+    "start": "tsc --watch",
+		"clean": "rm -rf lib",
+    "docs": "typedoc --excludePrivate --out ../../docs/swap src/index.ts --includeVersion --readme none",
+    "prepare": "run-s clean build",
+    "shell": "node -e \"$(< shell)\" -i --experimental-repl-await",
+    "test": "run-s test:unit test:lint test:build",
+    "test:build": "run-s build",
+    "test:lint": "eslint src",
+    "test:unit": "jest",
+    "test:watch": "jest --watch"
   },
   "devDependencies": {
     "@tsconfig/node12": "^1.0.7",
     "@types/bn.js": "^4.11.6",
-    "@types/bs58": "^4.0.1",
-    "@types/jest": "^24.9.1",
-    "@types/node": "^12.12.62"
-  }
+    "@types/jest": "^26.0.9",
+    "@typescript-eslint/eslint-plugin": "^4.6.0",
+    "@typescript-eslint/parser": "^4.6.0",
+    "babel-eslint": "^10.0.3",
+    "cross-env": "^7.0.2",
+    "eslint": "^7.6.0",
+    "eslint-config-prettier": "^6.11.0",
+    "jest": "^26.4.0",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.0.5",
+    "ts-jest": "^26.2.0",
+    "typedoc": "^0.20.36",
+    "typescript": "^4.0.5"
+  },
+  "files": [
+    "lib"
+  ],
+  "prettier": {
+    "singleQuote": true,
+    "trailingComma": "all"
+  },
+  "dependencies": {
+    "@project-serum/anchor": "^0.5.1-beta.2",
+    "@project-serum/serum": "^0.13.34",
+    "@solana/spl-token": "^0.1.3",
+    "@solana/spl-token-registry": "^0.2.68",
+    "@solana/web3.js": "^1.2.0",
+    "base64-js": "^1.5.1",
+    "bn.js": "^5.1.2"
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not op_mini all",
+    "maintained node versions"
+  ]
 }
